Added TODO, more stuff to gitignore, python initial setup, zshrc config

+++ make.sh
@@ -7,16 +7,30 @@
 # https://github.com/mathiasbynens/dotfiles
 
 
+# Ask for the administrator password upfront
+sudo -v
+
+# Keep-alive: update existing `sudo` time stamp until `.script` has finished
+while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
+
+
 ###############################################################################
 # install required things
 ###############################################################################
 # Requires XCode to be installed manually
 
+# oh-my-zsh
+curl -L http://install.ohmyz.sh | sh
+mv ~/.oh-my-zsh ~/.dotfiles/oh-my-zsh
+mv ~/.zshrc ~/.dotfiles/zshrc
+
+# python stuff
+bash python/install.sh
 
 ###############################################################################
 # set os defaults 
 ###############################################################################
-bash osx
+# bash osx
 
 
 ###############################################################################
@@ -24,13 +38,13 @@ bash osx
 ###############################################################################
 dir=~/.dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old              # old dotfiles backup directory
-files="bashrc vimrc zshrc oh-my-zsh gitignore"    # list of files/folders to symlink in homedir
+files="vimrc oh-my-zsh zshrc gitignore"    # list of files/folders to symlink in homedir
 
 
 ###############################################################################
 # create dotfiles_old in homedir
 ###############################################################################
-echo "Creating $olddir for backup of any existing dotfiles in ~"
+# echo "Creating $olddir for backup of any existing dotfiles in ~"
 mkdir -p $olddir
 echo "...done"
 
@@ -38,7 +52,7 @@ echo "...done"
 ###############################################################################
 # change to the dotfiles directory
 ###############################################################################
-echo "Changing to the $dir directory"
+# echo "Changing to the $dir directory"
 cd $dir
 echo "...done"
 
@@ -47,8 +61,12 @@ echo "...done"
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 ###############################################################################
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
+    # echo "Moving any existing dotfiles from ~ to $olddir"
+    if [ -e ".$file" ]
+    then
+        mv ~/.$file ~/dotfiles_old/
+    fi
+    # echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/.$file
 done
+
